requirements.txt
django
django-crispy-forms
django-filter
django-pure-pagination
rules
django-guardian